startup 代码中增加了函数 Iboot_GetResetAddr ，供 cpu.c中的 CPU_Reset等函数调用，对 C100 这种通过 boot0启动的CPU尤其重要。 cpu_peri_lcd.c:增加调用摄像头视频监控功能 sys_init.c:增加调用出++环境初始化所需要的 __CppRuntime 函数，定义符号：__dso_handle 由于C库中取消了符号 CN_TIMEOUT_FOREVER ，使用这个符号的代码都要加上 systime.h

+++ bsp/cpudrv/allwinner/f1cx00s/include/cpu_peri_io.h
@@ -14,10 +14,10 @@ typedef signed long long        loff_t;
 typedef signed long             register_t;
 typedef unsigned long           irq_flags_t;
 
-typedef unsigned long long      virtual_addr_t;
-typedef unsigned long long      virtual_size_t;
-typedef unsigned long long      physical_addr_t;
-typedef unsigned long long      physical_size_t;
+typedef unsigned int      virtual_addr_t;
+typedef unsigned int      virtual_size_t;
+typedef unsigned int      physical_addr_t;
+typedef unsigned int      physical_size_t;
 
 static   u8 read8(virtual_addr_t addr)
 {
